change fetch mechanism to json endpoint
new endpoint /fetchdata will do the fetching and return JSON with logs
and errors. The UI will run JS code to call the new endpoint. This will
show the load page immediately with a spinner to indicate that there is
something going on in the background until the data is present and logs
can be shown.

Signed-off-by: Aaron Lauterer <aaron@lauterer.at>