imap_extract: move handled emails to store_folder
We need to check the errors from insert_report as the UNIQUE contraint
one does seem to happen fairly often and is not a reason to skip to the
next email before moving the current one.

Mainly because for some reason, some reports are sent multiple times.

If we receive a different error, we skip before moving that email and it
will remain in the inbox.

Signed-off-by: Aaron Lauterer <aaron@lauterer.at>